refactor: separate Daily from TableBody
